// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/report-to-details-model.ts
export class ReportToDetailsModel {
    employeeId: string;
    firstName: string;
    surName: string;
    userName: string;
    email: string;
    employeeReportToId: string;
    reportToEmployeeId: string;
    reportToEmployeeFirstName: string;
    reportToEmployeeSurName: string;
    reportToEmployeeEmail: string;
    reportToEmployeeName: string;
    reportingMethodId: string;
    reportingMethod: string;
    otherText: string;
    activeFrom: string;
    activeTo: string;
    comments: string;
    timeStamp: any;
    totalCount: number;
    isArchived: boolean;
    reportingFrom: string;
    reportToUserId: string;
}